+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 <cite>Release contributors</cite>
 
-- 6 PR(s) by [Mykhailo Lytvyn](https://github.com/epam/sap-commerce-intellij-idea-plugin/pulls?q=milestone%3A2026.0.0+author%3Amlytvyn+is%3Apr)
+- 8 PR(s) by [Mykhailo Lytvyn](https://github.com/epam/sap-commerce-intellij-idea-plugin/pulls?q=milestone%3A2026.0.0+author%3Amlytvyn+is%3Apr)
 
 ### `CCv2` enhancements
 - Hide `delete a build` action for deployed builds [#1742](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1742)
@@ -11,6 +11,7 @@
 
 ### `ImpEx` enhancements
 - Optimized column delete action [#1745](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1745)
+- Inspection: improved detection of the unique type code & deployment table name [#1748](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1748)
 
 ### Other
 - Compatibility release for IDE 2026 [#1741](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1741)